requirements.txt
faiss-gpu
openai==1.35.13
python-dotenv==1.0.1
toml==0.10.2
torch==2.3.1
transformers==4.42.4
numpy
fastapi
uvicorn
pydantic
pytest
nltk
psycopg2-binary==2.9.9
black
flake8
streamlit
requests-mock
python-multipart